Implement Graph by HexGrid, adjust viewport

+++ core/src/com/gordonfromblumberg/games/core/common/graph/Edge.java
@@ -3,7 +3,7 @@ package com.gordonfromblumberg.games.core.common.graph;
 public interface Edge {
     Node getNode();
 
-    default int getWeight() {
-        return 1;
+    default float getWeight() {
+        return 1f;
     }
 }
